Move some tests to the right place.
