+++ extensions/blocks/publicize/twitter/test/utils.js
@@ -61,54 +61,3 @@ describe( 'contentAttributesChanged', () => {
 		expect( contentAttributesChanged( prevProps, props ) ).toBeFalsy();
 	} );
 } );
-
-describe( 'checkForTagsInContentAttributes', () => {
-	it( 'should return false for unsupported blocks', () => {
-		const props = {
-			attributes: {
-				content: '<strong>bar</strong>',
-			},
-			name: 'fake/block',
-		};
-		const tags = [ 'strong' ];
-
-		expect( checkForTagsInContentAttributes( props, tags ) ).toBeFalsy();
-	} );
-
-	it( 'should return false if the tags are not in the content', () => {
-		const props = {
-			attributes: {
-				content: '<i>bar</i>',
-			},
-			name: 'core/paragraph',
-		};
-		const tags = [ 'strong', 'em', 'b' ];
-
-		expect( checkForTagsInContentAttributes( props, tags ) ).toBeFalsy();
-	} );
-
-	it( 'should return true if the tags are in the content', () => {
-		const props = {
-			attributes: {
-				content: '<i>bar</i>',
-			},
-			name: 'core/paragraph',
-		};
-		const tags = [ 'strong', 'em', 'b', 'i' ];
-
-		expect( checkForTagsInContentAttributes( props, tags ) ).toBeTruthy();
-	} );
-
-	it( 'should return true if the tags are in only one attribute', () => {
-		const props = {
-			attributes: {
-				value: '<s>bar</s>',
-				citation: '<i>bar</i>',
-			},
-			name: 'core/quote',
-		};
-		const tags = [ 'strong', 'em', 'b', 'i' ];
-
-		expect( checkForTagsInContentAttributes( props, tags ) ).toBeTruthy();
-	} );
-} );
